erropr

+++ ms-catalogo/src/main/java/com/example/mscatalogo/entity/Producto.java
@@ -12,14 +12,14 @@ public class Producto {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer id;
     private String fechaVenta;
-    private int cantidadProductos;
+    private Integer cantidadProductos;
     private String metodoPago;
     private String nombreCliente;
     private String direccionCliente;
     private String correoCliente;
     private String telefonoCliente;
     private String productoVendido;
-    private double precioProducto;
+    private Integer precioProducto;
 
     @ManyToOne
     @JoinColumn(name = "catalogo_id")
